+++ client/src/components/List/List.stories.js
@@ -5,8 +5,57 @@ export default {
   title: "List",
   component: List,
 };
+const toDos = [
+  {
+    id: 1,
+    title: "Angebot anfordern",
+    category: "Location",
+    completed: false,
+  },
+  {
+    id: 2,
+    title: "Dj anrufen",
+    category: "Musik",
+    completed: true,
+  },
+  {
+    id: 3,
+    title: "Kleid auswählen",
+    category: "Bekleidung",
+    completed: false,
+  },
+  {
+    id: 4,
+    title: "Menü abstimmen",
+    category: "Location",
+    completed: false,
+  },
+  {
+    id: 5,
+    title: "Kleid auswählen",
+    category: "Bekleidung",
+    completed: false,
+  },
+  {
+    id: 6,
+    title: "Menü abstimmen",
+    category: "Location",
+    completed: false,
+  },
+  {
+    id: 7,
+    title: "Kleid auswählen",
+    category: "Bekleidung",
+    completed: false,
+  },
+  {
+    id: 8,
+    title: "Menü abstimmen",
+    category: "Location",
+    completed: false,
+  },
+];
 
-const Template = (args) => <List {...args} />;
+export const ListAllItems = () => <List items={toDos} />;
 
-export const ListAllItems = Template.bind({});
-ListAllItems.args = <List />;
+export const Empty = () => <List items={null} />;
